+++ CHANGELOG.md
@@ -259,3 +259,4 @@
 - 2024-01-20: Valid follow up, descend, tz country, campaign, +5
 - 2024-01-21: Builder chat, lead, discovery, bettercontact, +4
 - 2024-01-21: Add geo, lead, campaigns, LICENSE, +6 types
+- 2024-01-22: Cache conf, chat, lead, campaign, +5